Restructuring project directory
