+++ src/index.ts
@@ -1,7 +1,7 @@
 export { Matrix } from "./matrix";
 export { Vector } from "./vector";
 export { Scalar } from "./scalar";
-export { Evaluable, isEvaluable, Constant, isConstant, Variable, isVariable, Expression, isExpression, Operator } from "./definitions";
-export { ExpressionBuilder } from "./expression";
-export { BinaryOperator, isBinaryOperator } from "./operators";
-export { sin, cos, tan, asin, acos, atan, log, ln, exp, UnaryOperator, isUnaryOperator } from "./unary";
+export { Evaluable, isEvaluable, Constant, isConstant, Variable, isVariable, Expression, isExpression, Operator } from "./core/definitions";
+export { ExpressionBuilder } from "./core/expression";
+export { BinaryOperator, isBinaryOperator } from "./core/operators/binary";
+export { sin, cos, tan, asin, acos, atan, log, ln, exp, UnaryOperator, isUnaryOperator } from "./core/operators/unary";
