+++ tsconfig.json
@@ -63,7 +63,7 @@
   "exclude": [
     "node_modules/",
     "tests/",
-    "build",
-    "@types"
+    "build/",
+    "@types/"
   ],
 }
